Create an utils file for tests and clean code
